refactor: 조금 더 코드를 type safe 하게 변경
